# Server/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Beacon API
  version: 1.0.0
  description: |
    Two contracts:

      1. Workflows - CRUD over the `workflow` table.
      2. Workflow files - upload to Neon Object Storage, applying structural
         transformations *before* the object is written.

    Storage layout (bucket `data`):

        data/
          {workflow_id}/
            {filename}               each transformed data file
            _log.json                config_metadata, keyed by filename
            _struct_updates.json     live_updates, keyed by filename

    There is exactly ONE `_log.json` and ONE `_struct_updates.json` per
    workflow folder. Both accumulate: a second upload adds an entry under
    `files` rather than replacing the first file's.
servers:
  - url: /v1
...
paths:
...
  /workflows:
    post:
      tags: [workflows]
      operationId: createWorkflow
      summary: Create a workflow
      description: |
        Backs the "Create New Workflow" form. `state` defaults to `new` when
        omitted; `tag` is optional and may be null.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowCreate"
            examples:
              minimal:
                summary: Only the required field
                value:
                  workflow_name: Q3 Revenue Ingest
              full:
                summary: All fields supplied
                value:
                  workflow_name: Q3 Revenue Ingest
                  state: new
                  tag: finance
      responses:
        "201":
          description: Created.
          headers:
            Location:
              description: URI of the new workflow.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationFailed"

    get:
      tags: [workflows]
      operationId: listWorkflows
      summary: List workflows
      parameters:
        - name: state
          in: query
          description: Filter by exact state.
          schema:
            $ref: "#/components/schemas/WorkflowState"
        - name: tag
          in: query
          description: Filter by exact tag.
          schema:
            type: string
        - name: q
          in: query
          description: Case-insensitive substring match on workflow_name.
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          description: Opaque cursor from a previous response's next_cursor.
          schema:
            type: string
      responses:
        "200":
          description: A page of workflows.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Workflow"
                  next_cursor:
                    type: [string, "null"]
                    description: Pass as `cursor` for the next page; null on the last page.
        "400":
          $ref: "#/components/responses/BadRequest"
  /workflows/{workflow_id}:
    parameters:
      - $ref: "#/components/parameters/WorkflowId"

    get:
      tags: [workflows]
      operationId: getWorkflow
      summary: Fetch one workflow
      responses:
        "200":
          description: Found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [workflows]
      operationId: updateWorkflow
      summary: Partially update a workflow
      description: |
        Only supplied fields change. `tag` accepts an explicit null to clear
        it. `id` is immutable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowUpdate"
            examples:
              advanceState:
                summary: Move the workflow out of `new`
                value:
                  state: configured
              clearTag:
                summary: Remove the tag
                value:
                  tag: null
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"

    delete:
      tags: [workflows]
      operationId: deleteWorkflow
      summary: Delete a workflow
      description: |
        Deletes the row. Stored objects under `data/{workflow_id}/` are removed
        only when `purge_objects=true`; otherwise a workflow that still owns
        objects is rejected with 409 so files are never orphaned silently.
      parameters:
        - name: purge_objects
          in: query
          description: Also delete every object under the workflow's storage prefix.
          schema:
            type: boolean
            default: false
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Workflow still has stored objects and purge_objects was false.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  # ----------------------------------------------------------------------
  # Contract 2 :: Object storage
  # ----------------------------------------------------------------------
...
components:
  parameters:
    WorkflowId:
      name: workflow_id
      in: path
      required: true
      description: Workflow UUID. Doubles as the storage folder under bucket `data`.
      schema:
        type: string
        format: uuid
...
  schemas:
    WorkflowState:
      type: string
      description: |
        Workflow lifecycle state. `new` is the value set at creation. The set is
        intentionally open - the column is `text`, so adding a value needs an
        API allowlist change only, never a migration.
      default: new
      examples: [new, configured, running, complete, failed]
    Workflow:
      type: object
      description: A row of the `workflow` table.
      required: [id, workflow_name, state, tag]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        workflow_name:
          type: string
          minLength: 1
          maxLength: 200
        state:
          $ref: "#/components/schemas/WorkflowState"
        tag:
          type: [string, "null"]
          maxLength: 100
    WorkflowCreate:
      type: object
      required: [workflow_name]
      additionalProperties: false
      properties:
        workflow_name:
          type: string
          minLength: 1
          maxLength: 200
          description: Trimmed before storage; must not be blank once trimmed.
        state:
          allOf:
            - $ref: "#/components/schemas/WorkflowState"
          description: Defaults to `new` when omitted.
        tag:
          type: [string, "null"]
          maxLength: 100
          description: Optional.
    WorkflowUpdate:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        workflow_name:
          type: string
          minLength: 1
          maxLength: 200
        state:
          $ref: "#/components/schemas/WorkflowState"
        tag:
          type: [string, "null"]
          maxLength: 100
...
    Problem:
      type: object
      description: RFC 9457 problem detail.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: "about:blank"
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errors:
          type: array
          description: Field- or column-level causes.
          items:
            type: object
            properties:
              filename:
                type: string
              column:
                type: string
              pointer:
                type: string
                description: JSON Pointer into the request body.
              code:
                type: string
              message:
                type: string
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ValidationFailed:
      description: Body parsed but failed validation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: "https://beacon.api/problems/validation-failed"
            title: Validation failed
            status: 422
            errors:
              - pointer: /workflow_name
                code: blank
                message: workflow_name must not be blank.
...